COMP: fix warnings
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=16749
git-svn-id: http://svn.slicer.org/Slicer4/trunk@16749 3bd1e089-480b-0410-8dfb-8563597acbee

+++ QTModules/CropVolume/qSlicerCropVolumeModuleWidget.cxx
@@ -180,8 +180,8 @@ void qSlicerCropVolumeModuleWidget::onROIVisibilityChanged(){
     return;
 
   Q_D(qSlicerCropVolumeModuleWidget);
-  this->parametersNode->SetROIVisibility((int)d->VisibilityButton->isChecked());
-  this->parametersNode->GetROINode()->SetVisibility((int)d->VisibilityButton->isChecked());
+  this->parametersNode->SetROIVisibility(d->VisibilityButton->isChecked());
+  this->parametersNode->GetROINode()->SetVisibility(d->VisibilityButton->isChecked());
 }
 
 void qSlicerCropVolumeModuleWidget::onInterpolationModeChanged()
